make changes for better maintenance
